+++ chat-widget/src/components/webchatcontainerstateful/common/defaultStyles/defaultWebChatStyles.ts
@@ -54,5 +54,6 @@ export const defaultWebChatStyles: StyleOptions = {
         "<3": "❤️",
         "</3": "💔",
         "<\\3": "💔"
-    }
+    },
+    uploadMultiple: false
 };