+++ chat-widget/src/common/facades/FacadeChatSDK.ts
@@ -39,6 +39,7 @@ export class FacadeChatSDK {
     private isAuthenticated: boolean;
     private getAuthToken?: (authClientFunction?: string) => Promise<string | null>;
     private sdkMocked: boolean;
+    private disableReauthentication: boolean;
 
     public isSDKMocked(): boolean {
         return this.sdkMocked;
@@ -57,12 +58,13 @@ export class FacadeChatSDK {
         return !isNullOrEmptyString(this.token);
     }
 
-    constructor(input: IFacadeChatSDKInput) {
+    constructor(input: IFacadeChatSDKInput, disableReauthentication: boolean) {
         this.chatSDK = input.chatSDK;
         this.chatConfig = input.chatConfig;
         this.getAuthToken = input.getAuthToken;
         this.isAuthenticated = input.isAuthenticated;
         this.sdkMocked = input.isSDKMocked;
+        this.disableReauthentication = disableReauthentication;
     }
 
     //set default expiration to zero, for undefined or missed exp in jwt
@@ -130,6 +132,11 @@ export class FacadeChatSDK {
     }
 
     private async tokenRing(): Promise<PingResponse> {
+
+        if (this.disableReauthentication === true) {
+            // facade feature is disabled, so we are bypassing the re authentication and let it fail.
+            return { result: true, message: "Facade is disabled" };
+        }
 
         // this is needed for storybooks, specifically for reconnect pane which requires authentication bypass
         if (this.sdkMocked === true) {